toke.c: Allow \N{} handling fcn to be used elsewhere in core
This function will be used in regcomp.c in a later commit.  This commit
changes the function so that it is callable outside of toke.c.  It adds
a parameter and moves some code in new_constant to the wrapper function
so that these do not cause problems when called from outside toke.  And
it adds some assertions

(cherry picked from commit 25c7fb78daaad4f8ca1aaa9e19d6db99366388c7)

cperl: NULL charnames are still only asserted in perl5, cperl throws a
proper error.