toke.c: Change API of static function
This will be useful in future commits.  new_constant() is changed so
that if an extra parameter is not NULL, it sets it to point to an error
message instead of raising the message itself.  Thus its caller can
choose to handle errors itself.

(cherry picked from commit 164e423c70b80a4c3880f33561005d1fb870699c)