macro: using grammars (i.e. tokens) WIP
play with this syntax:

  macro <a:expr> "?=" <b:expr> {
    a = b if defined a;
  }
  macro "#ifdef" <expr> <s:stmtseq> {
    if ($Config{lc <expr>}) { s };
  }
  #ifdef USECPERL print "usecperl"

A macro def consists of series of tokens (existing terminals or non-terminals)
or strings, with the tokens having optional names (as in peg), and a block.
This way we can avoid the dreaded :infix, :postfix, ... attrs, and use a uniform
grammar syntax.
We can add new grammars, and optionally expose grammar objects to the user.
(a parser-combinator).

Enable yytname[] to get the token index for terminals and nonterminals
from a name at run-time, to be able to get all valid lexer tokens, and all
grammar rules.