Add global hash to handle \p{user-defined}
A global hash has to be specially handled.  The keys can't be shared,
and all the SVs stored into it must be in its thread.  This commit adds
the hash, and initialization, and macros for context change, but doesn't
use them.  The code to deal with this is entirely confined to regcomp.c.

(cherry picked from commit dd52e3cc434f4c6a495379f06a99d35da217eecb)