locale.c: Add detection of Turkic UTF-8 locales
When switching into a new locale, after it is decided this is a UTF-8
locale, the code now also checks for if the locale is a specialized
Turkic one, which has a couple of slightly modified casing change rules.
If so, it sets a flag indicating this.

The code that has been added in previous commits in this series check if
that flag is set when they are actually paying attention to the
background locale, and if so behave according to Unicode Turkic rules.

(cherry picked from commit 30d8090de81085bd3dff00c83a7ab6d3ff8dfc8d)