sv_utf8_upgrade_flags_grow(): Alloc extra byte if empty
People may call this expecting that the 'extra' parameter is on top of
whatever is in there.  If something is in there, that already includes a
NUL, but if nothing is in there, for safety, add a byte to the request.

(cherry picked from commit e2e3bb6ab2dc831e06e3f552bfafd7c66ad52179)