Add more checking for regnode offset overflowing
This is part of the ongoing failures in [perl #133921].

The bottom line cause is that there are generally 16 bits available for
the address of the next regnode.  On very large patterns, this may not
be enough.  When that happens, a long jump is used instead.

What previous commits have done is to insert tests in a loop to detect
that overflow isn't going to occur.  But it turns out that there are
other places where such overflow could occur.  The real solution should
be to detect overflow in the base level routine that would otherwise get
things wrong.  This entails making that routine be able to return
failure.  It turns out that another function is used under DEBUGGING, so
that one must be changed as well.  And the calls where it is possible
for this to overflow are changed to look for failure return and proceed
appropriately, which is to set a flag that we need to use long jumps,
and restart the parse.

(cherry picked from commit bf848a12528ab1e63a2f20da532eda498adbdca6)

cperl: f55b7b26c4afa4768435a9da9f0798829bbf41f7 added NN to regnode_offset,
but this is not a pointer and fails on modern compilers. This wrong decl
is still there.
cperl had this fixed in 41a5d5e3d1f831032e6fe2960c6efc6f8d8d9dc0