regexec.c: restrict match to substring in regmatch()
This begins to add the capability to regmatch() to not look beyond a
certain point in the string in order to find a match.  But this is
currently set to the end of the entire string so it shouldn't affect any
current outcomes.

(cherry picked from commit 255b1a66c426b857aeced6ded35298de6c4ad950)