Revert "win32: define HAS_BUILTIN_EXPECT on MinGW"
This reverts commit 69aa5ebd1f8de0b4ea51faeba005dbcb734e0bef.

RT #13360

This patch was tickling a bug in the compiler on that platform, the
details of which have not yet been diagnosed. Since the original commit
was just a minor optimisation, its been agreed to revert for now.

(cherry picked from commit abd494f123f7d413dd85748f8b52e08508476fee)